move utils, add separate pool addresses
